Mana, Material Libraries
Mana needs to be fixed. It decreases and than immediately increases when playing a card.

+++ Assets/Events/GenericEvents/UIUpdateEvent.cs
@@ -4,10 +4,10 @@ using UnityEngine;
 public class UIUpdateEvent : IGenericEvent
 {
     public int pHealth;
-    public int pMana;
+    public float pMana;
     public int pID;
 
-    public UIUpdateEvent(int pHealth, int pMana, int pID)
+    public UIUpdateEvent(int pHealth, float pMana, int pID)
     {
         this.pHealth = pHealth;
         this.pMana = pMana;
